requirements.txt
fastapi~=0.115.5
uvicorn
bcrypt~=4.2.1
sqlalchemy~=2.0.36
pydantic~=2.10.1
email-validator

pyshacl~=0.29.0
rdflib~=7.1.1
PyJWT~=2.10.0
starlette~=0.41.3
PyYAML
python-multipart
filelock
pytz
requests
tzdata
pyinstaller
datalad
git-annex
PyYAML
python-dateutil
jinja2
aiofiles
psycopg2-binary
# npm install --save-dev sass sass-loader
# npm install dayjs # added for datepicker
# pip install dump-things-service


